requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
imbalanced-learn>=0.8.0
streamlit>=1.13.0
requests>=2.26.0
pillow>=9.0.0
pickle-mixin>=1.0.2
data-golf>=0.5.1 